+++ src/core/request/response/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2021-??-??)
+
+#### :boom: Breaking Change
+
+* Now to freeze and clone response data will be used `core/object/proxy-readonly` and `core/object/proxy-clone` if it's possible
+
 ## v3.72.0 (2021-12-10)
 
 #### :rocket: New Feature